stickerbook/requirements.txt
opencv-python>=4.9
ultralytics
mobile-sam @ git+https://github.com/ChaoningZhang/MobileSAM.git
torch>=2.1
numpy>=1.26
pyyaml>=6.0
